ENH: Update Travis and AppVeyor config files with the latest tweaks from devel.
Also, the Qt 5.4 install in Travis OSX has been redone from scratch to dodge problems with Homebrew. To summarize, the precompiled Homebrew version for Yosemite is downloaded, extracted to the Homebrew Cellar, and the paths in binaries and libs are updated like Homebrew does. So, essentially we are emulating Homebrew to avoid fucking with it to get the desired version of Qt.